Fixed Round loop issue and accounting for draws
I was having trouble implementing rounds counts only when there was
a player or computerscore and not for when there were draws. I fixed the
issue when I implemented a while loop accounting for manual rounds
to increment and creating an else statement for draws without recalling
the game function again, which I was doing before and causing the game
to call itself twice within the loop.